Function calendrical_calculations::chinese_based::days_in_month

pub fn days_in_month<C: ChineseBased>(
    month: u8,
    new_year: RataDie,
    prev_new_moon: Option<RataDie>,
) -> (u8, RataDie)

Returns the number of days in the given month after the given new_year. Also returns the RataDie of the new moon beginning the next month.