Function calendrical_calculations::chinese_based::fixed_mid_year_from_year

pub fn fixed_mid_year_from_year<C: ChineseBased>(elapsed_years: i32) -> RataDie

Get a RataDie in the middle of a year.

This is not necessarily meant for direct use in calculations; rather, it is useful for getting a RataDie guaranteed to be in a given year as input for other calculations like calculating the leap month in a year.

Based on functions from Calendrical Calculations by Reingold & Dershowitz Lisp reference code: https://github.com/EdReingold/calendar-code2/blob/main/calendar.l#L5469-L5475