Function calendrical_calculations::chinese_based::month_structure_for_year

pub fn month_structure_for_year<C: ChineseBased>(
    new_year: RataDie,
    next_new_year: RataDie,
) -> ([bool; 13], Option<NonZeroU8>)

Returns the length of each month in the year, as well as a leap month index (1-indexed) if any.

Month lengths are stored as true for 30-day, false for 29-day. In the case of no leap months, month 13 will have value false.