calendrical_calculations::chinese_based

Function new_moon_before

pub(crate) fn new_moon_before<C: ChineseBased>(moment: Moment) -> RataDie

The fixed date in standard time at the observation location of the previous new moon before a given Moment.

Based on functions from Calendrical Calculations by Reingold & Dershowitz. Lisp reference code: https://github.com/EdReingold/calendar-code2/blob/main/calendar.l#L5318-L5327