calendrical_calculations::chinese_based

Function new_year_on_or_before_fixed_date

pub(crate) fn new_year_on_or_before_fixed_date<C: ChineseBased>(
    date: RataDie,
    prior_solstice: RataDie,
) -> (RataDie, RataDie)

Get the fixed date of the nearest Lunar New Year on or before a given fixed date. This function also returns the solstice following a given date for optimization (see #3743).

To call this function you must precompute the value of the prior solstice, which is the result of winter_solstice_on_or_before

Based on functions from Calendrical Calculations by Reingold & Dershowitz. Lisp reference code: https://github.com/EdReingold/calendar-code2/blob/main/calendar.l#L5396-L5405