calendrical_calculations

Module julian

The Julian calendar

Constants

DAYS_IN_4_YEAR_CYCLE 🔒

DAYS_IN_YEAR 🔒

JULIAN_EPOCH 🔒

Functions

day_before_year

Calculates the day before Jan 1 of year.

days_before_month

The number of days in this year before this month starts

easter

Calculates the date of Easter in the given year

fixed_from_julian

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/1ee51ecfaae6f856b0d7de3e36e9042100b4f424/calendar.l#L1689-L1709

fixed_from_julian_book_version

Get a fixed date from the ymd of a Julian date.

is_leap_year

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/1ee51ecfaae6f856b0d7de3e36e9042100b4f424/calendar.l#L1684-L1687

julian_from_fixed

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/1ee51ecfaae6f856b0d7de3e36e9042100b4f424/calendar.l#L1711-L1738

year_day

Calculates the month/day from the 1-based day of the year

year_from_fixed 🔒

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/1ee51ecfaae6f856b0d7de3e36e9042100b4f424/calendar.l#L1191-L1217