Crate calendrical_calculations

Calendrical calculations

This crate implements algorithms from Calendrical Calculations by Reingold & Dershowitz, Cambridge University Press, 4th edition (2018) as needed by ICU4X.

Most of these algorithms can be found as lisp code in the book or on GithHub.

The primary purpose of this crate is use by ICU4X, however if non-ICU4X users need this we are happy to add more structure to this crate as needed.

Modules

chinese_based

Chinese-like lunar calendars (Chinese, Dangi)

coptic

The Coptic calendar

ethiopian

The ethiopian calendar

hebrew

The Hebrew calendar

hebrew_keviyah

Alternate, more efficient structures for working with the Hebrew Calendar using the keviyah and Four Gates system

helpers

Additional math helpers

islamic

Various islamic lunar calendars

iso

The ISO calendar (also usable as Gregorian)

julian

The Julian calendar

persian

The persian calendar

rata_die

Representation of Rata Die (R.D.) dates, which are represented as the number of days since ISO date 0001-01-01.