calendrical_calculations
Module chinese_based

Chinese-like lunar calendars (Chinese, Dangi)

Structs

Chinese

A type implementing ChineseBased for the Chinese calendar

ChineseFromFixedResult

chinese_based_date_from_fixed returns extra things for use in caching

Dangi

A type implementing ChineseBased for the Dangi (Korean) calendar

YearBounds

Marks the bounds of a lunar year

Traits

ChineseBased

The trait ChineseBased is used by Chinese-based calendars to perform computations shared by such calendar. To do so, calendars should:

Functions

chinese_based_date_from_fixed

Get a chinese based date from a fixed date, with the related ISO year

days_in_month

Returns the number of days in the given month after the given new_year. Also returns the RataDie of the new moon beginning the next month.

days_in_prev_year

Given a new year, calculate the number of days in the previous year

days_in_provided_year

Calculated the numbers of days in the given year

days_until_month

Given the new year and a month/day pair, calculate the number of days until the first day of the given month

fixed_mid_year_from_year

Get a RataDie in the middle of a year.

get_leap_month_from_new_year

Given that new_year is the first day of a leap year, find which month in the year is a leap month.

is_leap_year

Whether this year is a leap year

last_month_day_in_year

The last month and day in this year

month_days

Returns the number of days in the given (year, month).

month_structure_for_year

Returns the length of each month in the year, as well as a leap month index (1-indexed) if any.