Module calendrical_calculations::chinese_based

Chinese-like lunar calendars (Chinese, Dangi)

Structs

Chinese

A type implementing ChineseBased for the Chinese calendar

ChineseFromFixedResult

chinese_based_date_from_fixed returns extra things for use in caching

Dangi

A type implementing ChineseBased for the Dangi (Korean) calendar

YearBounds

Marks the bounds of a lunar year

Traits

ChineseBased

The trait ChineseBased is used by Chinese-based calendars to perform computations shared by such calendar. To do so, calendars should:

Functions

chinese_based_date_from_fixed

Get a chinese based date from a fixed date, with the related ISO year

days_in_month

Returns the number of days in the given month after the given new_year. Also returns the RataDie of the new moon beginning the next month.

days_in_provided_year

Calculated the numbers of days in the given year

days_until_month

Given the new year and a month/day pair, calculate the number of days until the first day of the given month

fixed_mid_year_from_year

Get a RataDie in the middle of a year; this is not necessarily meant for direct use in calculations; rather, it is useful for getting a RataDie guaranteed to be in a given year as input for other calculations like calculating the leap month in a year.

get_leap_month_from_new_year

Given that new_year is the first day of a leap year, find which month in the year is a leap month. Since the first month in which there are no major solar terms is a leap month, this function cycles through months until it finds the leap month, then returns the number of that month. This function assumes the date passed in is in a leap year and tests to ensure this is the case in debug mode by asserting that no more than thirteen months are analyzed.

is_leap_year

Whether this year is a leap year

last_month_day_in_year

The last month and day in this year

month_days

Returns the number of days in the given (year, month). In the Chinese calendar, months start at each new moon, so this function finds the number of days between the new moon at the beginning of the given month and the new moon at the beginning of the next month.