Module calendrical_calculations::hebrew

The Hebrew calendar

Structs

BookHebrew

Biblical Hebrew dates. The months are reckoned a bit strangely, with the new year occurring on Tishri (as in the civil calendar) but the months being numbered in a different order

Constants

ADAR

The biblical month number used for the month of Adar (and Adar I)

ADARII

The biblical month number used for the month of Adar II

AV

The biblical month number used for the month of Av

ELUL

The biblical month number used for the month of Elul

IYYAR

The biblical month number used for the month of Iyyar

KISLEV

The biblical month number used for the month of Kislev

MARHESHVAN

The biblical month number used for the month of Marheshvan

NISAN

The biblical month number used for the month of Nisan

SHEVAT

The biblical month number used for the month of Shevat

SIVAN

The biblical month number used for the month of Sivan

TAMMUZ

The biblical month number used for the month of Tammuz

TEVET

The biblical month number used for the month of Tevet

TISHRI

The biblical month number used for the month of Tishri