calendrical_calculations

Module islamic

Various islamic lunar calendars

Structs

CivilIslamicMarker

Marker type for civil islamic calendar, for use with IslamicBasedMarker

ObservationalIslamicMarker

Marker type for observational islamic calendar, for use with IslamicBasedMarker

SaudiIslamicMarker

Marker type for Saudi islamic calendar, for use with IslamicBasedMarker

TabularIslamicMarker

Marker type for observational islamic calendar, for use with IslamicBasedMarker

Constants

CAIRO 🔒

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/main/calendar.l#L6898

FIXED_ISLAMIC_EPOCH_FRIDAY 🔒

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/main/calendar.l#L2066

FIXED_ISLAMIC_EPOCH_THURSDAY 🔒

Traits

IslamicBasedMarker

Common abstraction over islamic-style calendars

Functions

adjusted_saudi_criterion 🔒

fixed_from_islamic_civil

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/main/calendar.l#L2076

fixed_from_islamic_observational

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/main/calendar.l#L6904

fixed_from_islamic_tabular

Lisp code reference:https: //github.com/EdReingold/calendar-code2/blob/main/calendar.l#L2076

fixed_from_saudi_islamic

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/main/calendar.l#L6981

islamic_civil_from_fixed

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/main/calendar.l#L2090

islamic_tabular_from_fixed

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/main/calendar.l#L2090

observational_islamic_from_fixed

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/1ee51ecfaae6f856b0d7de3e36e9042100b4f424/calendar.l#L6983-L6995

observational_islamic_month_days

The number of days in a month for the observational islamic calendar

saudi_criterion 🔒

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/main/calendar.l#L6957

saudi_islamic_from_fixed

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/main/calendar.l#L6996

saudi_islamic_month_days

The number of days in a month for the Saudi (Umm Al-Qura) calendar

saudi_new_month_on_or_before

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/main/calendar.l#L6966