calendrical_calculations

Module islamic

Various islamic lunar calendars

Structs

Location

A Location on the Earth given as a latitude, longitude, elevation, and standard time zone. Latitude is given in degrees from -90 to 90, longitude in degrees from -180 to 180, elevation in meters, and zone as a UTC offset in fractional days (ex. UTC+1 would have zone = 1.0 / 24.0)

Constants

CAIRO

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/main/calendar.l#L6898

ISLAMIC_EPOCH_FRIDAY

Different islamic calendars use different epochs (Thursday vs Friday) due to disagreement on the exact date of Mohammed’s migration to Mecca. Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/main/calendar.l#L2066

ISLAMIC_EPOCH_THURSDAY

Different islamic calendars use different epochs (Thursday vs Friday) due to disagreement on the exact date of Mohammed’s migration to Mecca. Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/main/calendar.l#L2066

MEAN_YEAR_LENGTH

The average length of an Islamic year, equal to 12 moon cycles

MECCA

The location of Mecca; used for Islamic calendar calculations.

Functions

fixed_from_observational_islamic

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/main/calendar.l#L6904

fixed_from_saudi_islamic

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/main/calendar.l#L6981

fixed_from_tabular_islamic

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/main/calendar.l#L2076

observational_islamic_from_fixed

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/1ee51ecfaae6f856b0d7de3e36e9042100b4f424/calendar.l#L6983-L6995

observational_islamic_month_days

The number of days in a month for the observational islamic calendar

saudi_islamic_from_fixed

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/main/calendar.l#L6996

saudi_islamic_month_days

The number of days in a month for the Saudi (Umm Al-Qura) calendar

saudi_new_month_on_or_before

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/main/calendar.l#L6966

tabular_islamic_from_fixed

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/main/calendar.l#L2090