calendrical_calculations

Module julian

The Julian calendar

Functions

easter

Calculates the date of Easter in the given year

fixed_from_julian

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/1ee51ecfaae6f856b0d7de3e36e9042100b4f424/calendar.l#L1689-L1709

fixed_from_julian_book_version

Get a fixed date from the ymd of a Julian date.

is_leap_year

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/1ee51ecfaae6f856b0d7de3e36e9042100b4f424/calendar.l#L1684-L1687

julian_from_fixed

Lisp code reference: https://github.com/EdReingold/calendar-code2/blob/1ee51ecfaae6f856b0d7de3e36e9042100b4f424/calendar.l#L1711-L1738